% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindTables.R
\name{bindTables}
\alias{bindTables}
\title{Bind compatible data tables that have a similar name}
\usage{
bindTables(x)
}
\arguments{
\item{x}{a string pattern that is also included in all the data frames that have to be binded.}
}
\value{
All binded data frames that include in their name the string pattern given as input.
}
\description{
The \code{bindCompatibleTables} function will take a string pattern as input and look for all variables in the Global Environment that include that pattern. Then it will try to \code{rbind} all those variables.
}
\details{
It's important to have all data frames following the same naming pattern and avoid including that naming pattern in the names of other variables.
}
\examples{
\dontrun{
bindedDF <- bindTables("stringPattern")
}


}
