% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMissingDates.R
\name{findMissingDates}
\alias{findMissingDates}
\title{Find missing dates between oldest available date and yesterday in a vector.}
\usage{
findMissingDates(x)
}
\arguments{
\item{suite}{Character vector containing dates in YYYY-MM-DD format.}
}
\value{
A vector of missing dates in lubridate format.
}
\description{
This function will take a character vector containing dates and convert it to 'ymd' lubridate format. Then it will compare to an hypothetical vector containing all dates between the oldest date and yesterday.
}
\examples{
\dontrun{
ReportsIDs <- findMissingDates(x)
}
}
