% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{Provide a summary of in-memory objects per size}
\usage{
lsos(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE)
}
\value{
A summary of data frame like objects stored in-memory.
}
\description{
Provide a summary of in-memory objects per size
}
\details{
Useful summary of in-memory objects per size. Credit goes to Dirk Eddelbuettel.
}
\examples{
\dontrun{
lsos()
}

}
