% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelations.R
\name{MakeCorrelations}
\alias{MakeCorrelations}
\title{Add together two numbers.}
\usage{
MakeCorrelations(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{

}
