% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportsRetriever.R
\name{reportsRetriever}
\alias{reportsRetriever}
\title{Retrieve all reports stored as output of reportsGenerator function and consolidate them.}
\usage{
reportsRetriever(dataFrameReports)
}
\arguments{
\item{dataFrameReports}{This is the output from reportsGenerator function. It MUST contain a column titled: ReportID}
}
\value{
A data frame containing all the consolidated reports defined by the reportsGenerator function.
}
\description{
Retrieve all reports stored as output of reportsGenerator function and consolidate them.
}
\details{
It is recommended to break the input data frame in chunks of 50 rows in order to prevent memory issues if the reports are too large. Otherwise the server or local computer might run out of memory.
}
\examples{
\dontrun{
visitorActivity <- reportsRetriever(dataFrameReports)
}


}
