% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemaAlignment.R
\name{schemaAlignment}
\alias{schemaAlignment}
\title{schemaAlignment}
\usage{
schemaAlignment(x, y)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a similar data frame that might include additions like new rows, new columns or changed column names and/or positions.}
}
\value{
A new data frame that will include all columns that match 100% by name or by content. It follows the column arrangement of dataset #1 as closely as possible. But it only keeps the content of the second dataset.
}
\description{
Identify similarity between 2 datasets and generate a new one that follows as much as possible the schema available in the first dataset while keeping the info available in the second dataset.
}
\examples{
\dontrun{
z <- schemaAlignment(x,y)
}
}
