% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data based on randomized component of existing data.}
\usage{
simulateData()
}
\arguments{
\item{dateBegin}{This is the initial date for the data table}

\item{dateEnd}{This is the end date for the data table}

\item{averageDailyContacts}{This is the average amount of desired records to be included}

\item{numberContacts}{This is the amount of contacts (email accounts) to extract from a list of emails available.}

\item{Contacts}{This is a vector containing all available email accounts from which we want to use some to generate randomized valid data.}

\item{metrics}{This is a vector containing a string of metrics (columns) to simulate in the data table}

\item{metricSamples}{This is a list of vectors that include a set of valid values for each metric or a function to generate them. E.g. \code{(stri_rand_strings(100,6,"[A-Z]"))}}
}
\value{
A data frame containing simulated data.
}
\description{
Simulate data based on randomized component of existing data.
}
\examples{
\dontrun{
simulatedDF <- simulateData(dateBegin,dateEnd,numberContacts,contacts,metrics, metricSamples)
}

}
