% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveEmptyCols.R
\name{RemoveEmptyCols}
\alias{RemoveEmptyCols}
\title{Remove irrelevant columns from data frame}
\usage{
RemoveEmptyCols(x)
}
\arguments{
\item{x}{This is a data frame.}
}
\value{
The data frame used as an input without empty columns according to rules mentioned above.
}
\description{
Remove from data frame those columns that contain either:

- only empty values ("")

- only zeros (0)

- only NAs (NA)
}
\examples{
\dontrun{
cleanDataFrame <- RemoveEmptyCols(dataFrame)
}


}
